Update patch set 1
Patch Set 1:

I see the unit tests are failing in %check due to some missing BR, but that is not preventing the package from building because we are ignoring the results.

Could we fix this so we can enforce tests and catch errors like this in the future?

Patch-set: 1
Label: Workflow=0