Update patch set 1
Patch Set 1:

This seems to be caused by the latest u-c update for cliff (to 2.10.0). I've been doing some research, and [1] caused it.

Magnum inits DriverCommandManager using None as a namespace parameter [2]. With the cliff change, the _load_commands() method skips any loading, see [3], so this is causing the unit test error, which btw is also seen in the upstream gate.

I'm not sure if setting namespace as None is valid, so this could be fixed on the Magnum side or on the cliff side (if it's expected to be valid).

[1] - https://review.openstack.org/522380
[2] - https://github.com/openstack/magnum/blob/master/magnum/cmd/driver_manage.py#L87-L93
[3] - https://github.com/openstack/cliff/blob/2.10.0/cliff/commandmanager.py#L55-L56

Patch-set: 1